requirements.txt
Flask==3.0.0
google-generativeai==0.3.1
python-dotenv==1.0.0
pytest==7.4.3
requests==2.31.0
